New inlining schema implemented:
small functions that should be inlined in libFirm are implemented in _t.h files
with a __ prefix.
Preprocessor magic is used to automatically inline these functions whenever a _t.h
file is included instead of a .h file.
Note that this magic did not work outside libFirm without accessing _t.h files.

[r3085]