Remove an unneeded Sync edge when inlining.
When inlining a function, cparser produces a Sync after the parameters
have been copied, which has an edge to the Call's memory input. If there is
at least one struct argument, this edge is not needed, and will not be
generated now.

Also, if there is only one parameter, the Sync itself is not generated.