Kill unneeded operands when killing Stores.
After killing a store (e.g. if it is dead), its Store value can be killed,
if no other node uses it. We can use the code in reduce_adr_usage for that,
which also works for non-pointers.

This commit adds a new function kill_and_reduce_usage, which works for Load
and Store nodes. It kills the node and all Loads on which only the killed
node depends.