remove (a+b)*c => a*c + b*c reassoc rule for now
This rule produces more nodes than before and is therefore very
dangerous. The original argument that localopt rules will revert this in
non-profitable cases is not good enough as in combination with CSE, etc.
not everything will get reverted.

In the future someone could write an analysis that can detect cases
where the rule will have a positive effect and reenable it for these
cases.