Implemented matrix simplification in execfreq.
Only blocks which have a backward jump will be part of the linear system,
because the execution frequencies of other blocks can be computed by simple
substition from the predecessors.

This speeds up the execfreq algorithm and finds a solution more often, but may
still have problems with some ill-conditioned matrices.