change AR flags
Previously we had an AR commandline that replaced old object files with newer
ones in the archive. This however fails when we remove files from the
project, because the previous objects in the archive were not removed
(and often caused missing symbol errors when linking later).
So now I remove the archive each time before creating it, to make up for
the speed loss use the "thin archive" option. This is not a good idea
for a release/installed version of libfirm, but for that we still use
the automake build anyway.