Rework setting some OS details
The idea here, is that you do not simply set a target operating system
or architecture, but have a setting for all things different between the
operating systems. Knowledge about the name and habits of the operating
system is left to the frontend.
(We may want to provide a small support library in the future to share
the triple parsing and OS -> settings code between multiple frontends).