Remove produces_sp register constraint
We can easily identify the outputs producing a stack pointer by looking
if the stack pointer register is assigned. Unfortunately this also
needed a new mechanism to tell the verifier that some registers are not
in SSA form (all the time) which needs a few extra lines in each
backend.