ia32: Introduce FRAMEOFFSET immediate type
This way we don't need a separate frame_entity field in the attribute
struct anymore. We can also use the immediate type to determine whether
we need to assign a fresh stackentity to a node.
This commit does not change the frame_use field, which is now a
duplicate way to indicate whether we need to assign a stackentity. This
currently requires a lot of asserts to ensure the two stay in sync. In
the future I'd like to use frame_use only for the special case when a
32bit or 64bit entity needs to be forced.