arm, be, ia32, sparc: Do not go on with peephole optimization of IncSP, when be_peephole_IncSP_IncSP() succeeded.
The current IncSP was removed in this case and we now would operate on its predecessor IncSP.
This is wrong, e.g. when asking for currently free registers.
This corrects backend/incsp_pop.c, which was broken on ia32.
It was wrongly assumed that a certain register was free and could be used to turn an IncSP into a Pop.