Mark node as visited before actually transforming it.
This unifies the behavior in debug and optimize build, since
be_set_transformed_node also marks the node as visited.
Furthermore, nobody should call be_is_transformed for the current node,
since this would mean we ran into a cycle (which is covered by an assert in the
debug build).