iropt: Remove ConvP(AddI(ConvP())) transform instead add AddI/SubI(ConvP()) transform
The involved conversion are all same size and won't create extra target
instructions. So we can safely perform the transformtion at the Add/Sub
level to get a normalizing effect. This also fixes a problem where an
AddP was constructed with the wrong offset_mode.

This fixes firm-testsuite/opt/fehler307.c