Refactor fallthrough checks in ia32 and amd64 emitters
Unify code related to "can fallthrough?" checks, which is implemented
using irn links. The refactored code now is almost identical in the two
backends, the differences being that A) amd64 doesn't have SwitchJmp and
B) in the ia32 backend *predecessor* block links are used whereas in the
amd64 backend *successor* block links are used.