be: Improve handling of Copy before Perm.
The old code handled the case of a Copy followed by a binary Perm, i.e. 'a -> b; a <-> b', in lower_perm_node().
This situation does not exists anymore, because push_through_perm() moves the Perm through the Copy.
Also it is not triggered either, if copies instead of swaps are used to lower a Perm.
Now handle this case already in push_through_perm().
This has the added advantage, that cases like 'a -> b; { a -> c, b -> a }' are handled, too.