lower: Handle creation of the libgcc entities directly in the double word lowerer.
Before each backend specified the same names (or a subset thereof) by itself.
Now also backends without explicit handling for several double word operations just use the libgcc functions.
This re-uses the implementation of the SPARC backend, because it handled a superset of all others.

Additionally the arm backend did not mangle the name.
This was not critical because typical arm systems have no mangling.