ir: Overhaul representation of ASM constraints in the IR.
Now the constraints are a mapping from the template position to the input/output positions instead of the other way round.
Also there is only one list of constraints instead of a separate one for input and output.
This simplifies handling quite a bit, in particular the numbering.
E.g. "=m" actually is an input (for the address) and this caused miscounting when the constraints were processed.
Also processing of 64 bit operands also lead to miscounting.

This fixes x86code/asm_test15.c and x86code/asm_test19.c