perl-scripts: fix build for make & cmake
perl changed the behaviour that @INC does not include ./ anymore,
makefiles however used relative directories, so ./ was added, which is
fine. However, cmake uses a absolut directory, which results in the
execution of the path ./<absolut path> which breaks the compilation with
cmake. Adding ./ in the Makefile is enough for fixing this issue.