beuses: Make get_next_use() work when 'from' is the last instruction in a block.
This is necessary when jump instructions also produce normal values, e.g. asm goto.

Before it would get the next node in the schedule, i.e. the block.
But the block does not have a sensible time step.

Now instead of skipping to the next instruction just compare with a larger threshold.
Instructions in a block have consecutive time steps.