x86: Do not fold multiple Consts into an immediate.
The code only checked that each constant was small enough to encode into an immediate.
But adding multiple constants could overflow the 32 bit immediate.
On ia32 this is no issue, because registers are 32 bit wide, but on AMD64 this could lead to wrong results.
Usually this only happens when local optimizations are deactivated, otherwise multiple Consts would be combined into one anyway.

This fixes opt/add_consts.c on AMD64.