irio: Don't squash all imported frame types together
The previous attempt to reduce redundancy in imported
and then exported IR files in 26140a7 led to all imported
IR graphs referencing the same (global) frame type.

This commit partly reverts the aggressive deduplication of
26140a7, while still combining identifying initial types
added by `ir_init()` by their `type_nr`.

This would again result in many redundant frame types,
because the call to `new_ir_graph` generates a new
frame type each time it is called, which is not freed
when replaced by another frame type (the one we read
from the IR file to import) through a call to
`set_irg_frame_type`.

The solution is of course to free the initially allocated
frame type with `free_type`.