ia32: / ifconv: do not generate cmov constructs
In cases with larger if cascades the ifconversion tried to optimize
by building quite large blocks containing multiple mux nodes which will be lowered to
cmov instructions. This impacted the performance of the compiled program
heavily.

cmov generation can be enabled using the -mcmov option

fehler218 now fails (infinite loop in jump threading), but this bug
was only hidden and is not caused by disabling cmov generation.