Fix topological walker.
It could be the case that a node whose operand is a Phi was visited before
that Phi. This was because the visited flag was both used to limit the
recursion and to keep track of which nodes still needed to be visited
(i.e. have the walker called with them). This commit introduces a nodeset
to keep track of both separately.