be: Refine modelling of additional register pressure.
Now additional pressure is applied to the register pressure either before (positive value) or after (negative value) the instruction.

So far the value was applied to both the register pressure before and after the instruction.
This leads to overapproximation, e.g. for cltd (in: eax, out: edx).
When the input lives through then the register pressure after the instruction is 2, but +1 additional pressure unnecessarily increases it to 3.
Now the additional pressure is applied to either the register pressure before or after the instruction.
For cltd applying it only before the instruction is optimal, because the output can never be paired with the input.
Typical symptom was overspilling around cltd+idiv.

This still can overapproximate the actual register demand when in/out pairing depends on whether an input lives through.
E.g. in: eax+reg, out: edx.
Then 3 registers are needed when the reg input lives through. (additional pressure before 1)
But only 2 registers are needed when the reg input dies. (no additional pressure)

This fixes lit/overspill_cltd.c.